--- util/filterUtil.js
@@ -1,2 +1,0 @@
-// Eventually build out filtering based on subreddit, post, and comment
-
